+++ _source/LigaCancer/Data/Store/PatientStore.cs
@@ -227,7 +227,7 @@ namespace LigaCancer.Data.Store
                 {
                     patient.Family = new Family();
                 }
-                patient.Family.FamilyIncome += familyMember.MonthlyIncome;
+                patient.Family.FamilyIncome += (double)familyMember.MonthlyIncome;
                 patient.Family.FamilyMembers.Add(familyMember);
 
                 patient.Family.PerCapitaIncome = patient.Family.FamilyIncome / (patient.Family.FamilyMembers.Count() + 1);
